# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubernetes REST API",
    "version": "1.0.0",
    "description": "A simplified REST API for managing Kubernetes resources using client-go. Supports pods, deployments, services, namespaces, configmaps, persistent volume claims, nodes, and cluster info."
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
  "paths": {
...
    "/services": {
      "get": {
        "summary": "List services",
        "description": "Returns all services, optionally filtered by namespace.",
        "operationId": "listServices",
        "tags": ["Services"],
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of service summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ServiceSummary" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create a service",
        "description": "Creates a Kubernetes Service.",
        "operationId": "createService",
        "tags": ["Services"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateServiceRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceSummary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/services/{namespace}/{name}": {
      "get": {
        "summary": "Inspect a service",
        "description": "Returns the full Kubernetes Service object.",
        "operationId": "getService",
        "tags": ["Services"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "200": {
            "description": "Service object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceSummary" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a service",
        "operationId": "deleteService",
        "tags": ["Services"],
        "parameters": [
          { "$ref": "#/components/parameters/namespace" },
          { "$ref": "#/components/parameters/name" }
        ],
        "responses": {
          "204": { "description": "Service deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "namespace": {
        "name": "namespace",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Kubernetes namespace"
      },
      "name": {
        "name": "name",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource name"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or missing required fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Kubernetes API error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "pod not found" }
        }
      },
...
      "ServicePort": {
        "type": "object",
        "properties": {
          "name":        { "type": "string" },
          "protocol":    { "type": "string", "example": "TCP" },
          "port":        { "type": "integer" },
          "target_port": { "type": "string" },
          "node_port":   { "type": "integer" }
        }
      },
      "ServiceSummary": {
        "type": "object",
        "properties": {
          "name":         { "type": "string" },
          "namespace":    { "type": "string" },
          "type":         { "type": "string", "example": "ClusterIP" },
          "cluster_ip":   { "type": "string" },
          "external_ips": { "type": "array", "items": { "type": "string" } },
          "selector":     { "type": "object", "additionalProperties": { "type": "string" } },
          "labels":       { "type": "object", "additionalProperties": { "type": "string" } },
          "ports":        { "type": "array", "items": { "$ref": "#/components/schemas/ServicePort" } },
          "created_at":   { "type": "string", "format": "date-time" }
        }
      },
      "ServicePortSpec": {
        "type": "object",
        "required": ["port", "target_port"],
        "properties": {
          "name":        { "type": "string" },
          "protocol":    { "type": "string", "enum": ["TCP", "UDP"], "default": "TCP" },
          "port":        { "type": "integer" },
          "target_port": { "type": "integer" },
          "node_port":   { "type": "integer" }
        }
      },
      "CreateServiceRequest": {
        "type": "object",
        "required": ["name", "ports"],
        "properties": {
          "name":      { "type": "string" },
          "namespace": { "type": "string", "default": "default" },
          "type":      { "type": "string", "enum": ["ClusterIP", "NodePort", "LoadBalancer", "ExternalName"], "default": "ClusterIP" },
          "selector":  { "type": "object", "additionalProperties": { "type": "string" } },
          "labels":    { "type": "object", "additionalProperties": { "type": "string" } },
          "ports":     { "type": "array", "items": { "$ref": "#/components/schemas/ServicePortSpec" } }
        }
      },
...
    }
  },
...
}